+++ agrest-cayenne/src/test/java/io/agrest/cayenne/GET_IT.java
@@ -7,6 +7,7 @@ import io.agrest.cayenne.cayenne.main.E2;
 import io.agrest.cayenne.cayenne.main.E28;
 import io.agrest.cayenne.cayenne.main.E29;
 import io.agrest.cayenne.cayenne.main.E3;
+import io.agrest.cayenne.cayenne.main.E31;
 import io.agrest.cayenne.cayenne.main.E4;
 import io.agrest.cayenne.cayenne.main.E6;
 import io.agrest.cayenne.unit.AgCayenneTester;
@@ -40,7 +41,7 @@ public class GET_IT extends DbTest {
 
     @BQTestTool
     static final AgCayenneTester tester = tester(Resource.class)
-            .entities(E2.class, E3.class, E4.class, E6.class, E17.class, E19.class, E28.class)
+            .entities(E2.class, E3.class, E4.class, E6.class, E17.class, E19.class, E28.class, E31.class)
             .entitiesAndDependencies(E29.class)
             .build();
 
@@ -55,6 +56,18 @@ public class GET_IT extends DbTest {
                 .bodyEquals(1,
                         "{\"id\":1,\"cBoolean\":null,\"cDate\":null,\"cDecimal\":null,"
                                 + "\"cInt\":5,\"cTime\":null,\"cTimestamp\":null,\"cVarchar\":\"xxx\"}");
+    }
+
+    @Test
+    public void testIdCalledId() {
+
+        tester.e31().insertColumns("id", "name").values(5, "30").values(4, "31").exec();
+
+        tester.target("/e31")
+                .queryParam("sort", "name")
+                .get()
+                .wasOk()
+                .bodyEquals(2, "{\"id\":5,\"name\":\"30\"}", "{\"id\":4,\"name\":\"31\"}");
     }
 
     @Test
@@ -540,6 +553,12 @@ public class GET_IT extends DbTest {
             ids.put(E29.ID2PROP.getName(), id2);
 
             return AgJaxrs.select(E29.class, config).clientParams(uriInfo.getQueryParameters()).byId(ids).getOne();
+        }
+
+        @GET
+        @Path("e31")
+        public DataResponse<E31> getAllE31s(@Context UriInfo uriInfo) {
+            return AgJaxrs.select(E31.class, config).clientParams(uriInfo.getQueryParameters()).get();
         }
     }