+++ agrest-cayenne/src/test/java/io/agrest/cayenne/cayenne/main/E31.java
@@ -1,0 +1,9 @@
+package io.agrest.cayenne.cayenne.main;
+
+import io.agrest.cayenne.cayenne.main.auto._E31;
+
+public class E31 extends _E31 {
+
+    private static final long serialVersionUID = 1L;
+
+}